requirements.txt
openai
pyyaml
requests
anthropic
